Separate the original and getKeyMap
